+++ web/routes/web.php
@@ -1,6 +1,8 @@
 <?php
 
+use App\Exceptions\ShopifyProductCreatorException;
 use App\Lib\EnsureBilling;
+use App\Lib\ProductCreator;
 use App\Models\Session;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -10,6 +12,7 @@ use Illuminate\Support\Facades\Route;
 use Shopify\Auth\OAuth;
 use Shopify\Auth\Session as AuthSession;
 use Shopify\Clients\HttpHeaders;
+use Shopify\Clients\HttpResponse;
 use Shopify\Clients\Rest;
 use Shopify\Context;
 use Shopify\Utils;
@@ -67,7 +70,7 @@ Route::get('/api/auth', function (Request $request) {
     $installUrl = OAuth::begin(
         $shop,
         '/api/auth/callback',
-        true,
+        false,
         ['App\Lib\CookieHandler', 'saveShopifyCookie'],
     );
 
@@ -106,21 +109,7 @@ Route::get('/api/auth/callback', function (Request $request) {
     return redirect($redirectUrl);
 });
 
-Route::post('/api/graphql', function (Request $request) {
-    $response = Utils::graphqlProxy($request->header(), $request->cookie(), $request->getContent());
-
-    $xHeaders = array_filter(
-        $response->getHeaders(),
-        function ($key) {
-            return str_starts_with($key, 'X') || str_starts_with($key, 'x');
-        },
-        ARRAY_FILTER_USE_KEY
-    );
-
-    return response($response->getDecodedBody(), $response->getStatusCode())->withHeaders($xHeaders);
-})->middleware('shopify.auth:online');
-
-Route::get('/api/products-count', function (Request $request) {
+Route::get('/api/products/count', function (Request $request) {
     /** @var AuthSession */
     $session = $request->get('shopifySession'); // Provided by the shopify.auth middleware, guaranteed to be active
 
@@ -128,7 +117,37 @@ Route::get('/api/products-count', function (Request $request) {
     $result = $client->get('products/count');
 
     return response($result->getDecodedBody());
-})->middleware('shopify.auth:online');
+})->middleware('shopify.auth');
+
+Route::get('/api/products/create', function (Request $request) {
+    /** @var AuthSession */
+    $session = $request->get('shopifySession'); // Provided by the shopify.auth middleware, guaranteed to be active
+
+    $success = $code = $error = null;
+    try {
+        ProductCreator::call($session, 5);
+        $success = true;
+        $code = 200;
+        $error = null;
+    } catch (\Exception $e) {
+        $success = false;
+
+        if ($e instanceof ShopifyProductCreatorException) {
+            $code = $e->response->getStatusCode();
+            $error = $e->response->getDecodedBody();
+            if (array_key_exists("errors", $error)) {
+                $error = $error["errors"];
+            }
+        } else {
+            $code = 500;
+            $error = $e->getMessage();
+        }
+
+        Log::error("Failed to create products: $error");
+    } finally {
+        return response()->json(["success" => $success, "error" => $error], $code);
+    }
+})->middleware('shopify.auth');
 
 Route::post('/api/webhooks', function (Request $request) {
     try {