+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release history
 
-1.0.0 (June 20, 2022)
-----------
+## Unreleased
 
-* New Shopify App Template for PHP, with React frontend, for use with CLI 3
+-   ⚠️ [Breaking] Use offline access tokens to interact with the Admin API, remove GraphQL proxy since it doesn't work with offline tokens. ([#258](https://github.com/Shopify/shopify-app-template-php/pull/258))
+
+## 1.0.0 (June 20, 2022)
+
+-   New Shopify App Template for PHP, with React frontend, for use with CLI 3